// frontend/src/app/shared/api/model/create-image201-response.ts
/**
 * Food Driver API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateImage201Response { 
  [key: string]: any | any;


    /**
     * Unikalny identyfikator
     */
    idImage: number;
}