// frontend/src/app/shared/api/model/create-me-restaurant-menu-item-request.ts
/**
 * Food Driver API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CreateMeRestaurantMenuItemRequest {
  [key: string]: any | any;

  name: string;
  description: string;
  price: string;
  /**
   * Unikalny identyfikator
   */
  idImage: number;
}